backend/engine/scraper.py
import requests
from bs4 import BeautifulSoup
import logging
import json

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger("Scraper")

class InternshipScraper:
    def __init__(self):
        self.headers = {
            "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36"
        }

    def scrape_from_source(self, url: str):
        """Generic scraper for internship listings."""
        try:
            logger.info(f"Scraping {url}...")
            # Note: This is a placeholder as real scraping depends on specific site layout
            # and may require Playwright for JS-heavy sites.
            response = requests.get(url, headers=self.headers, timeout=10)
            soup = BeautifulSoup(response.content, "html.parser")
            
            # Simulated extraction
            listings = []
            # Example logic for a hypothetical job board
            # jobs = soup.find_all("div", class_="job-card")
            # for job in jobs:
            #     listings.append({
            #         "role": job.find("h2").text,
            #         "company": job.find("span", class_="company").text,
            #         "location": job.find("div", class_="location").text,
            #     })
            
            return listings
        except Exception as e:
            logger.error(f"Error scraping {url}: {e}")
            return []

if __name__ == "__main__":
    # Test scraping
    scraper = InternshipScraper()
    # results = scraper.scrape_from_source("https://example-jobs.com")
    # print(json.dumps(results))
    print("Scraper ready.")
